+++ apps/app/src/app.controller.ts
@@ -15,19 +15,27 @@ export class AppController {
   constructor(private readonly appService: AppService) {}
 
   @Post("sign/vc/:type")
-  async signVC(
+  async signCredential(
     @Param("type", new ParseEnumPipe(SignType)) type: SignType,
     @Body() body: SignRequestDto | EncryptedPayloadDto
   ) {
-    return await this.appService.signVC(type, body);
+    return await this.appService.signCredential(type, body);
   }
 
   @Post("sign/vp/:type")
-  async signVP(
+  async signPresentation(
     @Param("type", new ParseEnumPipe(SignType)) type: SignType,
     @Body() body: PresentRequestDto | EncryptedPayloadDto
   ) {
-    return await this.appService.signVP(type, body);
+    return await this.appService.signPresentation(type, body);
+  }
+
+  @Post("sign/pop/:type")
+  async signProofOfPossession(
+    @Param("type", new ParseEnumPipe(SignType)) type: SignType,
+    @Body() body: PresentRequestDto | EncryptedPayloadDto,
+  ) {
+    return await this.appService.signProofOfPossession(type, body);
   }
 
   @Post("generate")